requirements.txt
PyQt6~=6.3.1
pyqt6-tools
opencv-python~=4.6.0.66
numpy~=1.22.4
pyexiv2==2.8.1
GitPython
PyYAML~=6.0
Sphinx
sphinx-rtd-theme
sympy